[ui] 개인정보 처리방침 화면 작성 (#282)
